ref(boot.go,plansrv): rename plansrv to clustersrv
