+++ clustersrv/constants.go
@@ -1,4 +1,4 @@
-package plansrv
+package clustersrv
 
 const (
 	DefaultHTTPPort = 8080
