+++ clustersrv/plan_commit_handler.go
@@ -1,4 +1,4 @@
-package plansrv
+package clustersrv
 
 import (
 	"net/http"
